deps refactor. (#31)
* Skip test for deprecated y_pred_test and y_proba_test.

The test for y_pred_test and y_proba_test has been marked to skip as these parameters are deprecated and ignored. This change ensures that the test suite does not fail due to tests that are no longer relevant.

Signed-off-by: DanielAvdar <66269169+DanielAvdar@users.noreply.github.com>

* Add test for imports and update CI configuration

Introduce a new unit test to verify successful imports from the `deepchecks.tabular` module. Update the CI pipeline to include this test, and refactor the `pyproject.toml` by commenting out unnecessary dependencies to streamline the build process.

Signed-off-by: DanielAvdar <66269169+DanielAvdar@users.noreply.github.com>

* Remove cache setting for Python setup in CI workflow

The cache setting for poetry in the setup-python action was removed. This change optimizes the workflow by eliminating redundant cache specification, as caching is not required at this stage. The CI process should continue to run smoothly without this configuration.

Signed-off-by: DanielAvdar <66269169+DanielAvdar@users.noreply.github.com>

* Update CI to run specific unittest module

Modify the CI configuration to run the `test_imports.py` module exclusively. This ensures that only relevant tests are executed during the CI process, optimizing test time and focusing on crucial import tests.

Signed-off-by: DanielAvdar <66269169+DanielAvdar@users.noreply.github.com>

* Comment out requests dependency in pyproject.toml

This change temporarily disables the requests dependency. The update may be part of a broader refactoring or to address compatibility issues. Re-enabling might be necessary if dependent features require this package.

Signed-off-by: DanielAvdar <66269169+DanielAvdar@users.noreply.github.com>

* Comment out ipython dependency in pyproject.toml

The ipython dependency is now commented out to potentially prevent conflicts or redundancy in the project dependencies. This change does not remove the line completely, allowing for easy reactivation if needed in the future.

Signed-off-by: DanielAvdar <66269169+DanielAvdar@users.noreply.github.com>

* Clean up dependencies in pyproject.toml

Removed outdated and commented-out dependencies to streamline and modernize the dependency management. This change aims to minimize clutter and ensure that only actively used and necessary packages are included, improving maintainability.

Signed-off-by: DanielAvdar <66269169+DanielAvdar@users.noreply.github.com>

* Relocate ipykernel and ipywidgets to dev dependencies

Moved ipykernel and ipywidgets from regular dependencies to development dependencies in the pyproject.toml file. This change reflects that these packages are primarily needed for development and testing, not for the main package functionality.

Signed-off-by: DanielAvdar <66269169+DanielAvdar@users.noreply.github.com>

* Add matplotlib as a project dependency

This commit updates the pyproject.toml file to include matplotlib with a minimum version of 3.0.1. This ensures that the necessary plotting library is available for visualization tasks within the project.

Signed-off-by: DanielAvdar <66269169+DanielAvdar@users.noreply.github.com>

* Add ipywidgets to project dependencies

This commit includes the ipywidgets package with a minimum version of 8.0.1 to the project dependencies in pyproject.toml. This addition is necessary to ensure compatibility with interactive widgets used in the project.

Signed-off-by: DanielAvdar <66269169+DanielAvdar@users.noreply.github.com>

* Add plotly dependency to pyproject.toml

This change introduces the plotly library with a minimum version of 5.0.1 to the project's dependencies. Adding plotly enhances the project's data visualization capabilities, allowing for more interactive and dynamic visual representations. Make sure to install the updated dependencies to utilize plotly features.

Signed-off-by: DanielAvdar <66269169+DanielAvdar@users.noreply.github.com>

* Reorganize dependency and dev-dependency sections in pyproject.toml

Move 'ipywidgets' to maintain consistency in dependency order, ensuring 'ipywidgets' appears only in dependencies and aligns with the formatting. Rearrange and add developer dependencies under the correct section to improve the project's maintainability.

Signed-off-by: DanielAvdar <66269169+DanielAvdar@users.noreply.github.com>

---------

Signed-off-by: DanielAvdar <66269169+DanielAvdar@users.noreply.github.com>